+++ eval/Makefile.include
@@ -30,14 +30,19 @@ EVAL_SRC := \
     $(EVAL_DIR)/defs.cpp \
     $(SBG_DIR)/sbg_algorithms.cpp \
     $(SBG_DIR)/sbg.cpp \
-    $(SBG_DIR)/af_pwmap.cpp \
+    $(SBG_DIR)/pwmap_fact.cpp \
     $(SBG_DIR)/pw_map.cpp \
-    $(SBG_DIR)/af_map.cpp \
+    $(SBG_DIR)/unord_pwmap.cpp \
+    $(SBG_DIR)/ord_pwmap.cpp \
+    $(SBG_DIR)/map_fact.cpp \
     $(SBG_DIR)/map.cpp \
     $(SBG_DIR)/multidim_lexp.cpp \
     $(SBG_DIR)/lexp.cpp \
-    $(SBG_DIR)/af_set.cpp \
+    $(SBG_DIR)/set_fact.cpp \
     $(SBG_DIR)/set.cpp \
+    $(SBG_DIR)/unord_set.cpp \
+    $(SBG_DIR)/ord_set.cpp \
+    $(SBG_DIR)/ord_unidim_dense_set.cpp \
     $(SBG_DIR)/multidim_inter.cpp \
     $(SBG_DIR)/interval.cpp \
     $(SBG_DIR)/rational.cpp \
