+++ eval/visitors/eval_sbg.hpp
@@ -32,7 +32,7 @@ namespace Eval {
 
 struct EvalSBG : public boost::static_visitor<LIB::SBG> {
   public:
-  EvalSBG(unsigned int nmbr_dims, const LIB::PWMapAF &fact, VarEnv &env);
+  EvalSBG(unsigned int nmbr_dims, const LIB::PWMapFact &fact, VarEnv &env);
 
   LIB::SBG operator()(AST::Natural v) const;
   LIB::SBG operator()(AST::Rational v) const;
@@ -60,7 +60,7 @@ struct EvalSBG : public boost::static_visitor<LIB::SBG> {
 
   private:
   unsigned int nmbr_dims_;
-  const LIB::PWMapAF &fact_;
+  const LIB::PWMapFact &fact_;
   mutable VarEnv env_;
 };
